出力をintにした

+++ perceptron.py
@@ -10,7 +10,7 @@ class Perceptron:
         if len(input_data) != len(self.weights):
             print("入力数が違います")
             raise IndexError
-        return np.sum(self.weights * input_data) > self.threashold
+        return int(np.sum(self.weights * input_data) > self.threashold)
 
 if __name__ == '__main__':
     test = Perceptron(np.array([1, 1]), 1)
